many bug fixes and changes
- new feature - selectively compile individual chapters with the
command line flag `-c`, for example to compile chapter 2 run `$
./bin/compile -c 2`, which result in a chapter2.pdf file in the builds
folder, complete with references, and a title pager. References to
other sections will not work however

- allows for proper ampersand escaping, so the writer does not need to
do that: you can simply write “H&E” and it will get processed properly

- fixed inline code

-  fixed table and figure placement so that instead of waiting until
the end of a section, they are now placed on the very next page after
where they appear. In other words, the compiler will encounter a
figure, and finish filling out the current page with text, and put the
figure on the very next page. This is also true for tables.

- allows for the rudimentary parsing of table captions. The paragraph
immediately following a table is considered a table caption if it
starts with `:` or `Table:`. Then, just like figure captions, you make
include a table ID for in-text references and a short caption for the
list of tables like so: “Table: this is a long caption {#tab:my-table
short=“this is a short caption”}

- fixed tables so that they actually render correctly